Provide a Haskell solution for Day 2
Every year, it's a crash-course in trying to remember how
to build a parser. This year, I decided to try
Text.ParserCombinators.ReadP, as it's much simpler to get started.

I took some liberties in assuming that each line was fully parsed
by my parser, or at least that a prefix of each line could
be parsed.