Rudimentary support for a single Python runner
Each day's puzzle is a module named dayXX.py. The runner
takes a integer day number and imports the correct module,
then runs that module's `main_a` and `main_b` methods to
find the two daily answers.

The runner assumes a single input file named dayXX.input, but
can take part-specific inputs with the --a and --b options.

Days 1 and 2 have not been retrofitted to be used with runner.