Completed Day 4, part a
My initial guess of 135 was too low. I thought it was
because I failed to account for a passport that ended the file,
but in this case the file didn't end with an empty line, so that
particular bug wasn't triggered.

Turns out I didn't read the instructions closely, and failed to
allow a passport missing a country ID.