Cleanup and refactoring for part b
The biggest change is converting a Position from a heading and
an pair of coordinates to just a pair of coordinates, and
introducign a new type to store the combination of a heading and a
position.

This also resurrests a previous idea of separating the action given by
a navigation pair from the direction. Part b gets rid of the idea
of ship facing in a given direction with the idea of it always moving
towards a "waypoint", which is an aboslute position without a heading.